feat(pattern): adding pattern to server fix(vector): using deque

+++ Ecs/Component/include/Acceleration.hpp
@@ -7,22 +7,22 @@
 namespace ECS {
 class Acceleration : public IComp {
 public:
-    explicit Acceleration(int x, int y) noexcept;
+    explicit Acceleration(float x, float y) noexcept;
     Acceleration(const Acceleration& acc) noexcept;
     explicit Acceleration(Acceleration&&) noexcept = delete;
 
     Acceleration& operator=(const Acceleration&) const noexcept = delete;
     Acceleration& operator=(Acceleration&&) const noexcept = delete;
 
-    [[nodiscard]] int getAcceleration_x() const noexcept;
-    [[nodiscard]] int getAcceleration_y() const noexcept;
-    void setAcceleration_x(int x) noexcept;
-    void setAcceleration_y(int y) noexcept;
+    [[nodiscard]] float getAcceleration_x() const noexcept;
+    [[nodiscard]] float getAcceleration_y() const noexcept;
+    void setAcceleration_x(float x) noexcept;
+    void setAcceleration_y(float y) noexcept;
     ~Acceleration() noexcept override = default;
 
 private:
-    int _x;
-    int _y;
+    float _x;
+    float _y;
 };
 };
